refactor: profile카드 컴포넌트분리 및 이동
